destroy: aws: move iam helper functions to their own file
